add ItemManager, update ItemDao

+++ Escape_Room/src/main/java/model/entities/Clue.java
@@ -9,7 +9,7 @@ public class Clue extends Item {
     private Thematic thematic;
     private String details;
 
-    public Clue(int id, String name, float price, int idRoom, Type type, Thematic thematic, String details) {
+    public Clue(int id, String name, double price, int idRoom, Type type, Thematic thematic, String details) {
         super(id, name, price, idRoom, type);
         this.thematic = thematic;
         this.details = details;
