+++ Escape_Room/src/main/java/model/entities/Decoration.java
@@ -8,7 +8,7 @@ public class Decoration extends Item {
 
     private MaterialType material;
 
-    public Decoration(int id, String name, float price, int idRoom, Type type, MaterialType material) {
+    public Decoration(int id, String name, double price, int idRoom, Type type, MaterialType material) {
         super(id, name, price, idRoom, type);
         this.material = material;
     }
