+++ Escape_Room/src/main/java/model/entities/Item.java
@@ -10,7 +10,7 @@ public abstract class Item {
     private int idRoom;
     private Type type;
 
-    public Item(int id, String name, float price, int idRoom, Type type) {
+    public Item(int id, String name, double price, int idRoom, Type type) {
         this.id = id;
         this.name = name;
         this.price = price;
